Factor out `TensorBase` implementations into separate module/file
